Fixed sidebar lint regressions
ref https://linear.app/ghost/issue/BER-3458

This separates the custom sidebar views hook from the component file to satisfy the Fast Refresh lint rule and tightens the hook test mocks so the admin lint job passes again.
