+++ apps/admin/src/layout/app-sidebar/nav-custom-views.tsx
@@ -1,29 +1,8 @@
-import { useMemo } from 'react';
 import { NavSavedViews } from './nav-saved-views';
-import { useSharedViews } from './shared-views';
-import { useEmberRouting } from '@/ember-bridge';
+import { useCustomSidebarViews } from './use-custom-sidebar-views';
 
 interface NavCustomViewsProps {
     route?: 'posts' | 'pages';
-}
-
-export function useCustomSidebarViews(route: 'posts' | 'pages' = 'posts') {
-    const routing = useEmberRouting();
-    const sharedViews = useSharedViews(route);
-
-    return useMemo(() => {
-        return sharedViews.map((view) => {
-            const to = routing.getRouteUrl(route, view.filter);
-
-            return {
-                key: to,
-                name: view.name,
-                to,
-                isActive: routing.isRouteActive(route, view.filter),
-                color: view.color
-            };
-        });
-    }, [route, routing, sharedViews]);
 }
 
 export function NavCustomViews({ route = 'posts' }: NavCustomViewsProps) {
